Login test
Altered the SQL communication class to accept another argument being the
type (register or login) and created a switch statement within it to
allow the correct code to be executed depending on what was passed
through.
Created a test of logging in that outputs the entered login data to a
toast, checking the data passes to the SQL communicator correctly. Also
began creating the dashboard class and xml sheet to be logged into.