Login test - Passing data between activities
Added a new results constructor class (DbComResults) and code to the
login selector switch statement case (in DbCom). This allow objects to
be passed as the return value from doInBackground to onPostExectute
methods.
The returned objects contain a string value for the toast, a string
value to detail the method (register or login) and a string value of the
logged in user (that can be passed to the dashboard activity).
Additionally, a switch/case statement has been added to the
onPostExecute method so multiple different actions can occur depending
on whether the user is logging in or registering.
Also the MySQLConnection class has been renamed to DbCom.