Server response
Added a BufferedReader/InputStreamReader to read the input stream and
then a string builder to build the response from the server. This
response is then used tell the onPostExecute method what to do. This
will then be expanded to respond differently if the server flags an
error, for example if the username a user is registering already exists.