Clean up & Error checking improvements
Set username and password fields to a maximum length of 12 characters.
Altered the username error checking to allow for usernames of a minimum
of 5 characters and passwords of a minimum of 10 characters.
Updated PHP script and database accordingly to use correct field names.
Made password EditText fields use the type password.
Altered variable and class names to maintain consistency throughout.