Login - New activity & Logged in user
The doInBackground method returns an additional string (loggedInUser) to
the onPostExecute method, taken from the user variable.
When the user has been authenticated fully, the onPostExecute method
starts a new activity passing the logged in user string through the
intent. The new activity then to display the logged in user string.