Postdate timezone
Altered the addpost PHP script to set the timezone and get the date and
time within the script, rather than relying on the MySQL database to do
so. This was due to timezone issues and permissions to alter the
database settings. This allows the correct date to be posted to the
database with each dashboard post.