Shared preferences
The logged in user is no longer passed via the intent to load the
dashboard activity. Instead it is added to shared preferences. The
shared preferences is then cleared when the logout button is pressed.
This will allow the login menu to be bypassed if the user did not log
out in their previous session.