Shared preferences - Keep logged in
Created a PHP script that checks the connection to the database.
OnStart runs checking if a username is still stored in shared
preferences; if there is, the script is then ran. This essentially means
the user is still logged in, so if the connection is successful it
allows access to the dashboard activity logging the user in
automatically. If there's a connection error, the loggedInUser shared
preference is removed so the user is logged out.
When logged in and the logout button is pressed it returns the user to
the login activity, removing the loggedInUser shared preference.