connectionPost & connectionGet
Altered the connection method to be called connectionPost and created a
new method called connectionGet which doesn't send data to the server,
only receives a response. This is what the doInBackground checkcon
selector looks up to as it doesn't need to send data to the server. This
only needs to know if the connection to the database is successful or
not.