Exception handling
Added exception handling in each catch statement of the doInBackground
async task method. This prevents the app from crashing on exceptions
such as the device not having an internet connection. Each catch
statement now returns the selector result so it continues to look up to
the onPostExecute method regardless. As well as this, it also returns an
error string response to allow the onPostExecute method to display the
corresponding toast.