Pulling posts from database
Created a PHP script that stores all posts from the database, JSON
encodes it and then sends it to the application. The application
currently reads this data, loops through and logs individual items from
each object. This was done to test how to access specific parts of each
object. Therefore allowing data to be correctly added to the arraylist
and displayed in the list view.