Receives response from server
The loadposts php script now also sends a response alongside the array
of posts. This required the data to be stored in a JSON object before
the posts were extracted to a JSON array and the response was extracted
to a string. The onPostExecute method for loading posts is now defined
based on the response from the server.