Fully working posts filter
The filter switch is now fully functional without requiring a posts
refresh as it uses an onCheckedChanged listener.
The asynctask passes the list of posts to the dashboard activity via an
interface each time it's reloaded. This then runs the list through the
newly created PostsFilter class the filters and returns the required
list. The listview is then updated in accordance to the switches state.