Cleaned up code & Fixed known errors
Cleaned code in all classes.
Fixed an error where the app would crash when trying to load posts
instantly after deleting one. This was solved by using an interface
method that sends a string back to the postsadapter to say whether the
post was successfully deleted or not. If it was successfully deleted
from the database, then it is deleted from the list and the listview is
updated. This means the entire list doesn't need to be re-downloaded .