Username case sensitivity
Altered the case sensitivity of the username field in both the login and
register activity. This allows a user to register a username in any case
and it will be converted to lowercase (eg. user may enter uSerNAMe but
it will be read and stored as username.) Additionally the same occurs
when logging in, typing both strings in the example above will result in
accessing the same account.
This just makes it simpler to use as they're not physically limited on
strings. Also a lot of existing apps do the same.
Some special characters have allowed for use in the password fields.