Cancellable async task & connection timeout
Set the async task to be cancellable via a button on the progress
dialog. This is the only method of cancelling a running asynctask now.
Also added a timeout on the httpconnections to cancel the execution
after a set amount of seconds in a scenario where the connection is too
slow.