Swipe layout delete button
Implemented a custom library to handle swipe gestures on list items,
allowing a delete button to be displayed to the user. This is only
displayed on posts made by the specific user.
Also added a floating action button to add new posts.
The redesign is part complete but still need to format the ist items and
other activity content.