Fixed listview alignment issue
Prior to this commit, if posts varied in length, the swipe layout for
that post would not match the post height. To resolve this issue, when
posts are added to the listview, the heights are readjusted
programatically, by getting the height of the post and updating the post
swipe layout height.